# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Freedom Foundry API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user-profiles:
    get:
      operationId: getUserProfiles
      tags: [users]
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of user profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createUserProfile
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user-profiles/{id}:
    patch:
      operationId: updateUserProfile
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    FreeObject:
      type: object
      additionalProperties: true
...
    UserProfile:
      type: object
      properties:
        id:
          type: number
        user_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        business_name:
          type: string
        website:
          type: string
        phone:
          type: string
        headshot_url:
          type: string
        account_type:
          type: string
        brand_power_moves_unlocked:
          type: boolean
        brand_power_moves_unlocked_at:
          type: string
        unlock_method:
          type: string
        active_program_id:
          type: string
        notes:
          type: string
        setup_status:
          $ref: "#/components/schemas/FreeObject"
        marketing_consent:
          type: boolean
        consent_date:
          type: string
        created_at:
          type: string
      required: [id, user_id, created_at]
    UserProfileInput:
      type: object
      properties:
        user_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        business_name:
          type: string
        website:
          type: string
        phone:
          type: string
        headshot_url:
          type: string
        account_type:
          type: string
        brand_power_moves_unlocked:
          type: boolean
        brand_power_moves_unlocked_at:
          type: string
        unlock_method:
          type: string
        active_program_id:
          type: string
        notes:
          type: string
        setup_status:
          $ref: "#/components/schemas/FreeObject"
        marketing_consent:
          type: boolean
        consent_date:
          type: string

    # ── Brand ─────────────────────────────────────────────────────────────────
...